+++ examples/objectServerExample/src/main/java/io/realm/examples/objectserver/LoginActivity.java
@@ -77,7 +77,7 @@ public class LoginActivity extends AppCompatActivity {
         String password = this.password.getText().toString();
 
         Credentials creds = Credentials.usernamePassword(username, password, createUser);
-        String authUrl = "http://" + MyApplication.OBJECT_SERVER_IP + ":8080/auth";
+        String authUrl = "http://" + BuildConfig.OBJECT_SERVER_IP + ":9080/auth";
         User.Callback callback = new User.Callback() {
             @Override
             public void onSuccess(User user) {