+++ gradle-plugin/src/main/groovy/io/realm/gradle/Realm.groovy
@@ -41,7 +41,8 @@ class Realm implements Plugin<Project> {
             throw new GradleException('Realm gradle plugin only supports android gradle plugin 1.5.0 or later.')
         }
 
-        project.extensions.create('realm', RealmPluginExtension)
+        def syncEnabledDefault = false
+        project.extensions.create('realm', RealmPluginExtension, project, syncEnabledDefault)
 
         def usesKotlinPlugin = project.plugins.findPlugin('kotlin-android') != null
         def usesAptPlugin = project.plugins.findPlugin('com.neenbedankt.android-apt') != null
@@ -53,6 +54,7 @@ class Realm implements Plugin<Project> {
         }
 
         project.android.registerTransform(new RealmTransformer(project))
+
         project.repositories.add(project.getRepositories().jcenter())
         project.dependencies.add("compile", "io.realm:realm-annotations:${Version.VERSION}")
         if (isKaptProject) {
@@ -64,20 +66,6 @@ class Realm implements Plugin<Project> {
             project.dependencies.add("androidTestApt", "io.realm:realm-annotations:${Version.VERSION}")
             project.dependencies.add("androidTestApt", "io.realm:realm-annotations-processor:${Version.VERSION}")
         }
-
-        // Using afterEvaluate is now deprecated so this callback is used instead
-        def compileDeps = project.getConfigurations().getByName("compile").getDependencies()
-        project.getGradle().addListener(new DependencyResolutionListener() {
-            @Override
-            void beforeResolve(ResolvableDependencies resolvableDependencies) {
-                def suffix = project.realm.syncEnabled?'-object-server':''
-                compileDeps.add(project.getDependencies().create("io.realm:realm-android-library${suffix}:${Version.VERSION}"))
-                project.getGradle().removeListener(this)
-            }
-
-            @Override
-            void afterResolve(ResolvableDependencies resolvableDependencies) {}
-        })
     }
 
     private static boolean isTransformAvailable() {