+++ realm/realm-library/src/main/java/io/realm/RealmObjectSchema.java
@@ -570,7 +570,7 @@ public final class RealmObjectSchema {
         if (function != null) {
             long size = table.size();
             for (long i = 0; i < size; i++) {
-                function.apply(new DynamicRealmObject(realm, table.getCheckedRow(i), false));
+                function.apply(new DynamicRealmObject(realm, table.getCheckedRow(i)));
             }
         }