+++ realm/realm-library/src/objectServer/java/io/realm/SyncManager.java
@@ -21,6 +21,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
+import io.realm.annotations.Beta;
 import io.realm.internal.Keep;
 import io.realm.internal.network.AuthenticationServer;
 import io.realm.internal.network.OkHttpAuthenticationServer;
@@ -29,6 +30,7 @@ import io.realm.internal.objectserver.SyncSession;
 import io.realm.log.RealmLog;
 
 /**
+ * @Beta
  * The SyncManager is the central controller for interacting with the Realm Object Server.
  * It handles the creation of {@link Session}s and it is possible to configure session defaults and the underlying
  * network client using this class.
@@ -40,6 +42,7 @@ import io.realm.log.RealmLog;
  *
  */
 @Keep
+@Beta
 public final class SyncManager {
 
     /**
@@ -65,6 +68,8 @@ public final class SyncManager {
                 case RECOVERABLE:
                     RealmLog.info(errorMsg);
                     break;
+                default:
+                    throw new IllegalArgumentException("Unsupported error category: " + error.getErrorCode().getCategory());
             }
         }
     };
@@ -203,6 +208,7 @@ public final class SyncManager {
     // This is called from SyncManager.cpp from the worker thread the Sync Client is running on
     // Right now Core doesn't send these errors to the proper session, so instead we need to notify all sessions
     // from here. This can be removed once better error propagation is implemented in Sync Core.
+    @SuppressWarnings("unused")
     private static void notifyErrorHandler(int errorCode, String errorMessage) {
         ObjectServerError error = new ObjectServerError(ErrorCode.fromInt(errorCode), errorMessage);
         for (SyncSession session : SessionStore.getAllSessions()) {