+++ realm/realm-library/src/objectServer/java/io/realm/internal/objectserver/SyncObjectServerFacade.java
@@ -1,8 +1,25 @@
-package io.realm.internal.objectserver;
+/*
+ * Copyright 2016 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
+package io.realm.internal.objectserver;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
+import android.content.IntentFilter;
+import android.net.ConnectivityManager;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -14,8 +31,9 @@ import io.realm.SyncManager;
 import io.realm.exceptions.RealmException;
 import io.realm.internal.Keep;
 import io.realm.internal.ObjectServerFacade;
+import io.realm.internal.network.NetworkStateReceiver;
 
-@SuppressWarnings("unused") // Used through reflection. See ObjectServerFacade
+@SuppressWarnings({"unused", "WeakerAccess"}) // Used through reflection. See ObjectServerFacade
 @Keep
 public class SyncObjectServerFacade extends ObjectServerFacade {
 
@@ -45,7 +63,10 @@ public class SyncObjectServerFacade extends ObjectServerFacade {
             throw new RealmException("Could not initialize the Realm Object Server", e);
         }
         if (applicationContext == null) {
-            applicationContext = context;    
+            applicationContext = context;
+
+            applicationContext.registerReceiver(new NetworkStateReceiver(),
+                    new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION));
         }
     }
 
@@ -88,7 +109,7 @@ public class SyncObjectServerFacade extends ObjectServerFacade {
             SyncConfiguration syncConfig = (SyncConfiguration) config;
             String rosServerUrl = syncConfig.getServerUrl().toString();
             String rosUserToken = syncConfig.getUser().getAccessToken();
-            return new String[] {rosServerUrl, rosUserToken};
+            return new String[]{rosServerUrl, rosUserToken};
         } else {
             return new String[2];
         }